Fix: eventPublisher에서 @TransactionalEventListener 동작하지 않는 현상 수정(saveMessage()에 @Transactional 추가), updateLastMessage() 트랜잭션 propagation 기본값으로 수정(time out 현상 해결), ChatRoomRepository query 오타 수정

+++ src/main/java/com/example/HonBam/recipeapi/entity/Recipe.java
@@ -15,7 +15,7 @@ public class Recipe {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "data_id")
-    private int dataId;
+    private Long dataId;
 
     @Column(name = "cocktail_name")
     private String cocktailName;
